why have QA commands when we can just have unit tests? Move e2f link checks, and settings config checks, into unit tests
